Adds fix for cassandra-driver 3.0.0
A major update to the python cassandra-driver introduced a change
to the return value of 'execute'.  The new return value (ResultSet)
is a paginated iterator, whereas the old value was a list.

Attempting to perform a length operation now fails and instead requires
all pages to be fetched to determine the number of results returned.